+++ pmd/src/site/markdown/changelog.md
@@ -2,3 +2,8 @@
 
 ## ???? ??, 2014 - 5.2.0:
 
+**Bugfixes:**
+
+* Fixed [bug 1171]: Specifying minimum priority from command line gives NPE
+
+[bug 1171]: https://sourceforge.net/p/pmd/bugs/1171/